refactor(schedule): ScheduleItemCard를 Header/Detail로 분리 - ScheduleItemCard 수정   - 기존 Header + Detail 통합   - isChecked 상태 관리 및 toggle 처리   - click 이벤트(edit 모드 시) 처리   - UI 구조 단순화, 유지보수 용이하게 개선
